requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
jinja2>=3.1.0
python-dotenv>=1.0.0
